Update integration with back-end new types
