+++ src/cashier_frontend/src/components/confirmation-drawer/confirmation-drawer-assets-section.tsx
@@ -2,7 +2,7 @@ import { ArrowUpDown, Info } from "lucide-react";
 import { FC } from "react";
 import { useTranslation } from "react-i18next";
 import { TransactionItem } from "@/components/transaction/transaction-item";
-import { IntentModel } from "@/services/types/refractor.intent.service.types";
+import { IntentModel } from "@/services/types/intent.service.types";
 
 type ConfirmationPopupAssetsSectionProps = {
     intents: IntentModel[];
