+++ src/cashier_frontend/src/components/confirmation-drawer/confirmation-drawer.hooks.ts
@@ -1,5 +1,5 @@
 import { INTENT_STATE, TASK } from "@/services/types/enum";
-import { IntentModel } from "@/services/types/refractor.intent.service.types";
+import { IntentModel } from "@/services/types/intent.service.types";
 import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
