+++ src/cashier_frontend/src/pages/edit/[id]/index.tsx
@@ -24,7 +24,7 @@ import { Identity } from "@dfinity/agent";
 import { toCanisterCallRequest } from "@/services/types/mapper/intent.service.mapper";
 import useToast from "@/hooks/useToast";
 import { getCashierError } from "@/services/errorProcess.service";
-import { ActionModel } from "@/services/types/refractor.action.service.types";
+import { ActionModel } from "@/services/types/action.service.types";
 import { useLinkDataQuery } from "@/hooks/useLinkDataQuery";
 import { LINK_TEMPLATE_DESCRIPTION_MESSAGE } from "@/constants/message";
 
@@ -79,7 +79,7 @@ export default function LinkPage({ initialStep = 0 }: { initialStep?: number }) 
             }
 
             if (action) {
-                console.log("added link action", action);
+                console.log("🚀 ~ useEffect ~ action:", action);
                 setLinkAction(action);
             }
