+++ src/cashier_frontend/src/services/types/action.service.types.ts
@@ -1,5 +1,5 @@
 import { ACTION_TYPE } from "./enum";
-import { IntentModel } from "./refractor.intent.service.types";
+import { IntentModel } from "./intent.service.types";
 
 export type ActionModel = {
     id: string;
